nova driver - only request bare server objects when listing IDs
The nova driver was requesting the default level of detail for Server
objects from openstacksdk and then only using the IDs. For a project
with 13 instances with up to 4 ports each, this was taking over a minute
to return and causing an Ironic consumer to time out. With bare objects
only, the same call took 14 seconds instead of 100 seconds.